+++ src/main/java/pl/asie/computronics/util/cipher/RSACalculationTask.java
@@ -3,7 +3,11 @@ package pl.asie.computronics.util.cipher;
 import pl.asie.lib.util.Base64;
 
 import java.math.BigInteger;
-import java.security.SecureRandom;
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+import java.security.interfaces.RSAPrivateCrtKey;
+import java.security.interfaces.RSAPrivateKey;
+import java.security.interfaces.RSAPublicKey;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -13,11 +17,6 @@ import java.util.concurrent.Callable;
  * @author Vexatos
  */
 public class RSACalculationTask implements Callable<ArrayList<Map<Integer, String>>> {
-
-	private static final BigInteger
-		ONE = BigInteger.ONE,
-		TWO = new BigInteger("2"),
-		SEVENTEEN = new BigInteger("17");
 
 	private int bitLength = 0;
 	private int p = 0;
@@ -48,19 +47,44 @@ public class RSACalculationTask implements Callable<ArrayList<Map<Integer, Strin
 		return this.createKeySet();
 	}
 
+	// Random key pair creation
+
 	private ArrayList<Map<Integer, String>> createKeySet() {
-		SecureRandom r = new SecureRandom();
-		return this.createKeySet(
-			new BigInteger(1024, 100, r),
-			new BigInteger(1024, 100, r));
+		return this.createKeySet(1024);
 	}
 
-	private ArrayList<Map<Integer, String>> createKeySet(int bitLength) {
-		SecureRandom r = new SecureRandom();
-		return this.createKeySet(
-			new BigInteger(bitLength, 100, r),
-			new BigInteger(bitLength, 100, r));
+	private static final ThreadLocal<KeyPairGenerator> gen = new ThreadLocals.LocalKeyPairGenerator();
+
+	private ArrayList<Map<Integer, String>> createKeySet(int keylength) {
+		KeyPairGenerator gen = RSACalculationTask.gen.get();
+		if(gen == null) {
+			return null;
+		}
+		gen.initialize(keylength);
+		KeyPair pair = gen.generateKeyPair();
+		if(pair == null || !(pair.getPublic() instanceof RSAPublicKey) || !(pair.getPrivate() instanceof RSAPrivateKey)) {
+			return null;
+		}
+		RSAPublicKey pubKey = (RSAPublicKey) pair.getPublic();
+		RSAPrivateCrtKey privKey = (RSAPrivateCrtKey) pair.getPrivate();
+		Map<Integer, String> publicKey = new LinkedHashMap<Integer, String>();
+		Map<Integer, String> privateKey = new LinkedHashMap<Integer, String>();
+		publicKey.put(1, Base64.encodeBytes(pubKey.getModulus().toByteArray()));
+		publicKey.put(2, Base64.encodeBytes(pubKey.getPublicExponent().toByteArray()));
+		privateKey.put(1, Base64.encodeBytes(privKey.getModulus().toByteArray()));
+		privateKey.put(2, Base64.encodeBytes(privKey.getPrivateExponent().toByteArray()));
+		ArrayList<Map<Integer, String>> list = new ArrayList<Map<Integer, String>>();
+		list.add(publicKey);
+		list.add(privateKey);
+		return list;
 	}
+
+	// Non-random key pair creation
+
+	private static final BigInteger
+		ONE = BigInteger.ONE,
+		TWO = new BigInteger("2"),
+		SEVENTEEN = new BigInteger("17");
 
 	private ArrayList<Map<Integer, String>> createKeySet(int p, int q) {
 		return this.createKeySet(
@@ -81,8 +105,10 @@ public class RSACalculationTask implements Callable<ArrayList<Map<Integer, Strin
 		Map<Integer, String> privateKey = new LinkedHashMap<Integer, String>();
 		publicKey.put(1, Base64.encodeBytes(n.toByteArray()));
 		publicKey.put(2, Base64.encodeBytes(d.toByteArray()));
+		publicKey.put(3, "prime");
 		privateKey.put(1, Base64.encodeBytes(n.toByteArray()));
 		privateKey.put(2, Base64.encodeBytes(e.toByteArray()));
+		privateKey.put(3, "prime");
 		ArrayList<Map<Integer, String>> list = new ArrayList<Map<Integer, String>>();
 		list.add(publicKey);
 		list.add(privateKey);